requirements.txt
aioboto3==13.0.1
aiobotocore==2.13.0
aiofiles==23.2.1
aiohttp==3.9.5
aioitertools==0.11.0
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.3.0
appdirs==1.4.4
attrs==23.2.0
beautifulsoup4==4.12.3
billiard==3.6.4.0
black==20.8b1
blinker==1.8.2
boto3==1.34.106
botocore==1.34.106
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crontab==1.0.1
dnspython==2.6.1
email_validator==2.1.1
fake-useragent==1.5.1
fastapi==0.111.0
fastapi-cli==0.0.4
flake8==7.0.0
freezegun==1.5.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
motor==3.4.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
odmantic==1.0.2
orjson==3.10.3
outcome==1.3.0.post0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
prometheus_client==0.20.0
prompt_toolkit==3.0.46
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pyflakes==3.2.0
Pygments==2.18.0
pymongo==4.7.2
PySocks==1.7.1
pytest==8.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.2
rich==13.7.1
s3transfer==0.10.1
selenium==4.21.0
setuptools==70.0.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
starlette==0.37.2
toml==0.10.2
trio==0.25.1
trio-websocket==0.11.1
typed_ast==1.5.5
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.23.2
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.13
webdriver-manager==4.0.1
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.19.2
